+++ Expanded Cheat Menu/Code/Homeless.lua
@@ -4,6 +4,12 @@
 
 local Translate = ChoGGi.ComFuncs.Translate
 
+-- now to get the devs to add this by default (Added the X so people realise it isn't offical)
+local Mods,rawset = Mods,rawset
+for id,mod in pairs(Mods) do
+	rawset(mod.env,"CurrentModId_X",id)
+end
+
 -- bugfix?
 -- got me, MapTools shouldn't be doing anything
 if not rawget(_G,"DroneDebug") then