+++ Expanded Cheat Menu/Changelog.md
@@ -10,6 +10,7 @@
 - Debug>Loading Screen Log: Be able to see the console log (and other dialogs) during the loading screen.
 - ECM>Game>Lightmodel>List Normal: Changes the list of lightmodels to use (night/day/etc).
 - ECM>Game>Lightmodel>List Disaster: Overrides List Normal.
+- Examine>Context menu (functions)>Function Results: Continually call a function while showing results in an examine dialog.
 
 ### Changed:
 - Mod Upload supports batch uploading (and screenshots).